* Delphi compatibility.
git-svn-id: https://mseide-msegui.svn.sourceforge.net/svnroot/mseide-msegui/trunk@3822 56ecf5fe-6917-0410-8ac6-ca8e178bccbf

+++ lib/common/kernel/msereal.pas
@@ -263,7 +263,7 @@ begin
     end;
    end;
   end;
-  result:= trystrtofloat(str1,value);
+  result:= trystrtofloat(str1,double(value));
  end;
 end;
